Seperated view models into their own file. Changed MobilePassiveData to latest version
