+++ iosViewBuilder/iosViewBuilder.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -15,8 +15,8 @@
         "repositoryURL": "https://github.com/Sage-Bionetworks/MobilePassiveData-SDK.git",
         "state": {
           "branch": null,
-          "revision": "1c271f053bd3be9e2f45ea4c1c3cced3e47304c1",
-          "version": "1.3.1"
+          "revision": "e614b0f469230cdec16446842a148a55fe3eb6d3",
+          "version": "1.3.2"
         }
       },
       {
